moved ErrorReportable to model
