+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/GoogleServicesDAO.scala
@@ -12,7 +12,9 @@ import spray.json.JsObject
 import scala.concurrent.Future
 import scala.util.Try
 
-abstract class GoogleServicesDAO(groupsPrefix: String) {
+abstract class GoogleServicesDAO(groupsPrefix: String) extends ErrorReportable {
+  val errorReportSource = ErrorReportSource("google")
+
   val CREATE_PROJECT_OPERATION = "create_project"
 
   val billingEmail: String
