+++ google/src/main/scala/org/broadinstitute/dsde/rawls/google/package.scala
@@ -1,0 +1,7 @@
+package org.broadinstitute.dsde.rawls
+
+import org.broadinstitute.dsde.rawls.model.ErrorReportSource
+
+package object google {
+  implicit val errorReportSource = ErrorReportSource("google")
+}
